fix: copy style image when training

+++ server/src/magic_style/images/views.py
@@ -13,6 +13,7 @@ from rpc import rpc_interface
 from utils.api import APIView, check
 from utils.constants.account import UserTypeEnum
 from utils.constants.images import StyleImageTypeEnum
+from utils.constants.rpc import TrainingModeTypeEnum
 from utils.shortcuts import rand_str, save_file, delete_file
 
 
@@ -96,15 +97,16 @@ class TrainingModeAPI(APIView):
     def post(self, request):
         operation = request.data['operation']
 
-        for_train_images = StyleImage.objects.filter(image_type=StyleImageTypeEnum.for_train)
-        StyleImage.objects.filter(image_type=StyleImageTypeEnum.trained).delete()
-        for image in for_train_images:
-            image_name = rand_str()
-            with open(os.path.join(settings.STYLE_IMAGE_PATH, image_name), 'wb') as f:
-                with open(os.path.join(settings.STYLE_IMAGE_FORTRAIN_PATH, image.now_name), 'rb') as fp:
-                    f.write(fp.read())
-            StyleImage.objects.create(upload_name=image.upload_name, now_name=image_name,
-                                      image_type=StyleImageTypeEnum.trained)
+        if operation == TrainingModeTypeEnum.start:
+            for_train_images = StyleImage.objects.filter(image_type=StyleImageTypeEnum.for_train)
+            StyleImage.objects.filter(image_type=StyleImageTypeEnum.trained).delete()
+            for image in for_train_images:
+                image_name = rand_str()
+                with open(os.path.join(settings.STYLE_IMAGE_PATH, image_name), 'wb') as f:
+                    with open(os.path.join(settings.STYLE_IMAGE_FORTRAIN_PATH, image.now_name), 'rb') as fp:
+                        f.write(fp.read())
+                StyleImage.objects.create(upload_name=image.upload_name, now_name=image_name,
+                                          image_type=StyleImageTypeEnum.trained)
 
         try:
             resp = rpc_interface.training_mode(operation)
